pipelines/sentinel1/requirements.txt
aws-lambda-powertools
stactools-sentinel1
